[feat] : 나의 게시물만 삭제가능 처리

+++ backend/src/main/java/com/ssafy/fullerting/exArticle/service/ExArticleService.java
@@ -319,6 +319,13 @@ public class ExArticleService {
 
         List<Image> image = imageRepository.findAllByExArticleId(ex_article_id);
 
+        UserResponse userResponse = userService.getUserInfo();
+        CustomUser user = userResponse.toEntity(userResponse);
+
+        if (article.getUser().getId() != user.getId()) {
+            throw new ExArticleException(ExArticleErrorCode.NOT_MINE);
+        }
+
         for (Image image1 : image) {
             imageRepository.delete(image1);
             amazonS3Service.deleteFile(image1.getImgStoreUrl()); //s3
@@ -370,7 +377,7 @@ public class ExArticleService {
     }
 
     public ExArticle modifyarticle(Long exArticleId,
-                                   UpdateArticleRequest updateArticleRequest, CustomUser customUser,List<MultipartFile> files) {
+                                   UpdateArticleRequest updateArticleRequest, CustomUser customUser, List<MultipartFile> files) {
         ExArticle article = exArticleRepository.findById(exArticleId).orElseThrow(() -> new ExArticleException
                 (ExArticleErrorCode.NOT_EXISTS));
 
@@ -419,9 +426,9 @@ public class ExArticleService {
 //                    .build());
 //        });
 
-        if ( !files.get(0).isEmpty()) {
+        if (!files.get(0).isEmpty()) {
 
-            log.info("size 가 0보다 크다" +  files.get(0));
+            log.info("size 가 0보다 크다" + files.get(0));
             //이미지 업로드
             S3ManyFilesResponse response = amazonS3Service.uploadFiles(files);
             //이미지 DB 저장
